+++ app-test.js
@@ -1,7 +1,0 @@
-﻿
-const greeting = "Hello"; 
-
-
-function test() {
-  return greeting;
-}